Update helm-charts repo with the master branch

+++ dojo/tool_product/views.py
@@ -1,19 +1,22 @@
 # #  product
 import logging
 from django.contrib import messages
-from django.contrib.auth.decorators import user_passes_test
+from django.core.exceptions import BadRequest
+from django.http import HttpResponseRedirect
 from django.urls import reverse
 from django.shortcuts import render, get_object_or_404
-from django.utils import timezone
-from dojo.forms import DeleteToolProductSettingsForm, ToolProductSettingsForm, NoteForm
+from django.utils.translation import gettext as _
+
+from dojo.forms import DeleteToolProductSettingsForm, ToolProductSettingsForm
 from dojo.models import Product, Tool_Product_Settings
-from django.http import HttpResponseRedirect
-from dojo.utils import add_breadcrumb, Product_Tab
+from dojo.utils import Product_Tab
+from dojo.authorization.authorization_decorators import user_is_authorized
+from dojo.authorization.roles_permissions import Permissions
 
 logger = logging.getLogger(__name__)
 
 
-@user_passes_test(lambda u: u.is_staff)
+@user_is_authorized(Product, Permissions.Product_Edit, 'pid')
 def new_tool_product(request, pid):
     prod = get_object_or_404(Product, id=pid)
     if request.method == 'POST':
@@ -27,13 +30,13 @@ def new_tool_product(request, pid):
             messages.add_message(
                 request,
                 messages.SUCCESS,
-                'Product Tool Configuration Successfully Created.',
+                _('Product Tool Configuration Successfully Created.'),
                 extra_tags='alert-success')
             return HttpResponseRedirect(
                 reverse('all_tool_product', args=(pid, )))
     else:
         tform = ToolProductSettingsForm()
-    product_tab = Product_Tab(pid, title="Tool Configurations", tab="settings")
+    product_tab = Product_Tab(prod, title=_("Tool Configurations"), tab="settings")
     return render(request, 'dojo/new_tool_product.html', {
         'tform': tform,
         'product_tab': product_tab,
@@ -41,11 +44,11 @@ def new_tool_product(request, pid):
     })
 
 
-@user_passes_test(lambda u: u.is_staff)
+@user_is_authorized(Product, Permissions.Product_Edit, 'pid')
 def all_tool_product(request, pid):
     prod = get_object_or_404(Product, id=pid)
     tools = Tool_Product_Settings.objects.filter(product=prod).order_by('name')
-    product_tab = Product_Tab(prod.id, title="Tool Configurations", tab="settings")
+    product_tab = Product_Tab(prod, title=_("Tool Configurations"), tab="settings")
     return render(request, 'dojo/view_tool_product_all.html', {
         'prod': prod,
         'tools': tools,
@@ -53,47 +56,12 @@ def all_tool_product(request, pid):
     })
 
 
-@user_passes_test(lambda u: u.is_staff)
-def view_tool_product(request, pid, ttid):
-    tool = Tool_Product_Settings.objects.get(pk=ttid)
-    notes = tool.notes.all()
-
-    if request.method == 'POST':
-        form = NoteForm(request.POST)
-        if form.is_valid():
-            new_note = form.save(commit=False)
-            new_note.author = request.user
-            new_note.date = timezone.now()
-            new_note.save()
-            tool.notes.add(new_note)
-            form = NoteForm()
-            # url = request.build_absolute_uri(reverse("view_test", args=(test.id,)))
-            # title="Test: %s on %s" % (test.test_type.name, test.engagement.product.name)
-            # process_notifications(request, new_note, url, title)
-            messages.add_message(
-                request,
-                messages.SUCCESS,
-                'Note added successfully.',
-                extra_tags='alert-success')
-    else:
-        form = NoteForm()
-
-    add_breadcrumb(
-        title="View Product Tool Configuration",
-        top_level=False,
-        request=request)
-
-    return render(request, 'dojo/view_tool_product.html', {
-        'tool': tool,
-        'notes': notes,
-        'form': form
-    })
-
-
-@user_passes_test(lambda u: u.is_staff)
+@user_is_authorized(Product, Permissions.Product_Edit, 'pid')
 def edit_tool_product(request, pid, ttid):
-    prod = get_object_or_404(Product, id=pid)
+    product = get_object_or_404(Product, id=pid)
     tool_product = Tool_Product_Settings.objects.get(pk=ttid)
+    if tool_product.product != product:
+        raise BadRequest(f'Product {pid} does not fit to product of Tool_Product {tool_product.product.id}')
 
     if request.method == 'POST':
         tform = ToolProductSettingsForm(request.POST, instance=tool_product)
@@ -102,51 +70,41 @@ def edit_tool_product(request, pid, ttid):
             messages.add_message(
                 request,
                 messages.SUCCESS,
-                'Tool Product Configuration Successfully Updated.',
+                _('Tool Product Configuration Successfully Updated.'),
                 extra_tags='alert-success')
             return HttpResponseRedirect(reverse('all_tool_product', args=(pid, )))
     else:
         tform = ToolProductSettingsForm(instance=tool_product)
 
-    product_tab = Product_Tab(pid, title="Edit Product Tool Configuration", tab="settings")
+    product_tab = Product_Tab(product, title=_("Edit Product Tool Configuration"), tab="settings")
     return render(request, 'dojo/edit_tool_product.html', {
         'tform': tform,
         'product_tab': product_tab
     })
 
 
-@user_passes_test(lambda u: u.is_staff)
+@user_is_authorized(Product, Permissions.Product_Edit, 'pid')
 def delete_tool_product(request, pid, ttid):
     tool_product = Tool_Product_Settings.objects.get(pk=ttid)
-    prod = get_object_or_404(Product, id=pid)
+    product = get_object_or_404(Product, id=pid)
+    if tool_product.product != product:
+        raise BadRequest(f'Product {pid} does not fit to product of Tool_Product {tool_product.product.id}')
+
     if request.method == 'POST':
-        tform = DeleteToolProductSettingsForm(
-            request.POST, instance=tool_product)
+        DeleteToolProductSettingsForm(request.POST, instance=tool_product)
         tool_product.delete()
         messages.add_message(
             request,
             messages.SUCCESS,
-            'Tool Product Successfully Deleted.',
+            _('Tool Product Successfully Deleted.'),
             extra_tags='alert-success')
         return HttpResponseRedirect(reverse('all_tool_product', args=(pid, )))
     else:
         tform = ToolProductSettingsForm(instance=tool_product)
 
-    product_tab = Product_Tab(pid, title="Delete Product Tool Configuration", tab="settings")
+    product_tab = Product_Tab(product, title=_("Delete Product Tool Configuration"), tab="settings")
 
     return render(request, 'dojo/delete_tool_product.html', {
         'tform': tform,
         'product_tab': product_tab
-    })
-
-
-@user_passes_test(lambda u: u.is_staff)
-def tool_product(request):
-    confs = Tool_Product_Settings.objects.all().order_by('name')
-    add_breadcrumb(
-        title="Tool Configuration List",
-        top_level=not len(request.GET),
-        request=request)
-    return render(request, 'dojo/tool_product.html', {
-        'confs': confs,
     })
